fix: refactor server/index.js (also note it's now named serverStartingPoint.js)
