+++ package.json
@@ -2,7 +2,7 @@
   "name": "boilerplate",
   "version": "0.1.0",
   "description": "A boilerplate for creating a new project",
-  "main": "server/index.js",
+  "main": "server/serverStartingPoint.js",
   "proxy": "http://localhost:8000",
   "dependencies": {
     "@testing-library/jest-dom": "^5.11.4",
@@ -20,10 +20,10 @@
     "web-vitals": "^1.0.1"
   },
   "scripts": {
-    "start": "cross-env NODE_ENV=production node server/index.js",
+    "start": "cross-env NODE_ENV=production node server/serverStartingPoint.js",
     "build": "cross-env NODE_ENV=production run-s react:test react:build",
     "dev:start": "cross-env NODE_ENV=development run-p react:start dev:server:start",
-    "dev:server:start": "nodemon server/index.js",
+    "dev:server:start": "nodemon server/serverStartingPoint.js",
     "react:start": "react-scripts start",
     "react:build": "react-scripts build",
     "react:test": "react-scripts test"
